feat: improve token browser UI and UX - Add token name display, standardize network boxes, add testnet toggle, improve selector height, refactor wizard components into separate files
